requirements.txt
ultralytics>=8.3.0
pandas>=2.1.0
pyyaml>=6.0.0

